requirements.txt
-e .
aws-cdk.core>=1.8.0
aws-cdk.aws_iam
aws-cdk.aws_s3
aws-cdk.aws_cloudfront
aws-cdk.aws_route53
aws-cdk.aws_route53_targets
aws-cdk.aws_certificatemanager
aws-cdk.aws_codebuild
aws-cdk.aws_codepipeline
aws-cdk.aws_codepipeline_actions
